Split controllers
- Switched to action and view controllers
