+++ app/Http/Controllers/View/Index.php
@@ -1,8 +1,9 @@
 <?php
 declare(strict_types=1);
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\View;
 
+use App\Http\Controllers\Controller;
 use App\Models\ShareToken;
 use Illuminate\Http\Request;
