+++ resources/views/game.blade.php
@@ -76,14 +76,14 @@
 
                         @if ($game['complete'] !== 1)
                         <div class="pb-3 pt-1">
-                            <form action="{{ route('game.delete', ['game_id' => $game['id']]) }}" method="POST" class="d-inline">@csrf <button type="submit" class="btn btn-sm btn-danger">Delete Game</button></form>
+                            <form action="{{ route('game.delete.action', ['game_id' => $game['id']]) }}" method="POST" class="d-inline">@csrf <button type="submit" class="btn btn-sm btn-danger">Delete Game</button></form>
                         </div>
                         <div class="pb-1">
                             <a href="{{ route('game.add-players.view', ['game_id' => $game['id']]) }}" class="btn btn-sm btn-primary">Add Additional Players</a>
                         </div>
                         <div>
-                            <form action="{{ route('game.complete', ['game_id' => $game['id']]) }}" method="POST" class="d-inline">@csrf <button type="submit" class="btn btn-sm btn-primary">Complete Game</button></form>
-                            <form action="{{ route('game.complete.play-again', ['game_id' => $game['id']]) }}" method="POST" class="d-inline">@csrf <button type="submit" class="btn btn-sm btn-primary">Complete Game & Play Again</button></form>
+                            <form action="{{ route('game.complete.action', ['game_id' => $game['id']]) }}" method="POST" class="d-inline">@csrf <button type="submit" class="btn btn-sm btn-primary">Complete Game</button></form>
+                            <form action="{{ route('game.complete.play-again.action', ['game_id' => $game['id']]) }}" method="POST" class="d-inline">@csrf <button type="submit" class="btn btn-sm btn-primary">Complete Game & Play Again</button></form>
                         </div>
                         @endif
                     </li>
