+++ resources/views/new-player.blade.php
@@ -16,7 +16,7 @@
         <div class="col-lg-8 mx-auto p-3 py-md-5">
 
             <main>
-                <form action="{{ route('player.create.process') }}" method="POST" class="col-12 col-md-4 col-lg-4 mx-auto p-2">
+                <form action="{{ route('player.create.action') }}" method="POST" class="col-12 col-md-4 col-lg-4 mx-auto p-2">
                     <div class="mb-3">
                         <h2>New Player</h2>
                         <p>Add a new player, they will be selectable as a player in all new games.</p>
